pyproject.toml
[project]
name = "astrollm"
version = "0.1.0"
description = "A domain-specialized Large Language Model for Astronomy & Astrophysics"
readme = "README.md"
requires-python = ">=3.11"
license = "Apache-2.0"

dependencies = [
    # Core ML
    "torch>=2.4.0",
    "transformers>=4.45.0",
    "peft>=0.13.0",
    "trl>=0.12.0",
    "accelerate>=1.0.0",
    "bitsandbytes>=0.44.0",
    "datasets>=3.0.0",
    "tokenizers>=0.20.0",

    # Experiment Tracking
    "wandb>=0.18.0",

    # Model Merging
    "mergekit>=0.0.5",

    # Data Pipeline
    "arxiv>=2.1.0",
    "pdfminer.six>=20231228",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "ftfy>=6.2.0",
    "langdetect>=1.0.9",
    "datasketch>=1.6.0",

    # Astronomy Tools
    "astropy>=6.0.0",
    "astroquery>=0.4.7",
    "ads>=0.12.6",

    # RAG
    "sentence-transformers>=3.0.0",
    "pgvector>=0.3.0",
    "psycopg2-binary>=2.9.0",

    # Evaluation
    "lm-eval>=0.4.0",
    "rouge-score>=0.1.2",
    "sacrebleu>=2.4.0",

    # Utilities
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.66.0",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "jsonlines>=4.0.0",
    "rich>=13.7.0",
    "typer>=0.12.0",
    "httpx>=0.27.0",
]

[dependency-groups]
# GPU-only deps — install on cloud training instances, not local macOS
training = [
    "unsloth[colab-new]>=2024.11",
    "flash-attn>=2.6.0",
    "xformers>=0.0.28",
    "auto-gptq>=0.7.0",
    "autoawq>=0.2.0",
]

[tool.uv]
# triton is a torch transitive dep that only has Linux wheels
override-dependencies = [
    "triton>=3.6.0; sys_platform == 'linux'",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
